% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{convert_repo_date}
\alias{convert_repo_date}
\title{Function to convert date from a specific format to 'ddmmyyyy'}
\usage{
convert_repo_date(date_input)
}
\description{
Function to convert date from a specific format to 'ddmmyyyy'
}
\keyword{internal}
