% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn_2024_v1.R
\docType{data}
\name{iucn_2024_v1}
\alias{iucn_2024_v1}
\title{IUCN Red List Dataset 2024 (v1)}
\format{
A tibble with 163,040 rows and 16 variables:
\describe{
\item{assessment_id}{Integer. Unique identifier for each IUCN assessment.}
\item{internal_taxon_id}{Integer. Internal ID for each taxon within the IUCN system.}
\item{scientific_name}{Character. The scientific name of the taxon.}
\item{kingdom_name}{Character. The kingdom to which the taxon belongs.}
\item{phylum_name}{Character. The phylum to which the taxon belongs.}
\item{order_name}{Character. The order of the taxon.}
\item{class_name}{Character. The class of the taxon.}
\item{family_name}{Character. The family to which the taxon belongs.}
\item{genus_name}{Character. The genus of the taxon.}
\item{species_name}{Character. The species name of the taxon.}
\item{authority}{Character. The taxonomic authority for the species, indicating the source of its scientific name.}
\item{redlist_category}{Character. The IUCN Red List category of the taxon, indicating its conservation status (e.g., "Near Threatened", "Critically Endangered").}
\item{redlist_criteria}{Character. The criteria used to assign the Red List category.}
\item{criteria_version}{Numeric. Version of the criteria used in the assessment, typically 3.1 or 2.3.}
\item{population_trend}{Character. The current trend in the population of the taxon (e.g., "Decreasing", "Stable").}
\item{scopes}{Character. Geographic scope of the assessment, such as "Global" or "Global & Europe".}
}
}
\usage{
iucn_2024_v1
}
\description{
A dataset containing the International Union for Conservation of
Nature (IUCN) Red List taxonomy and assessment data for 163,040 taxa,
as of 2024. The data includes information on taxonomic classification,
Red List Category, Red List Criteria, criteria version, and population
trends.
}
\keyword{datasets}
