% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{iucn_check_version}
\alias{iucn_check_version}
\title{Check if the packaged version of iucnr is up to date.}
\usage{
iucn_check_version(silent = FALSE)
}
\arguments{
\item{silent}{Raise a warning if the version is out of date.}
}
\description{
Check if the packaged version of iucnr is up to date.
}
