% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{iucnrdata_version}
\alias{iucnrdata_version}
\title{Functions to help check the iucnr data is okay and up to date.
View the version of iucnrdata available.}
\usage{
iucnrdata_version(long = TRUE)
}
\arguments{
\item{long}{Whether to return the version date with version number.}
}
\description{
Functions to help check the iucnr data is okay and up to date.
View the version of iucnrdata available.
}
